test: ensure checkByEmail() return false id email not found in db
