+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:56911/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:51732/","mongoDBName":"jest"}
