+++ src/infra/db/mongodb/index.ts
@@ -1,1 +1,2 @@
 export * from './mongo-helpers'
+export * from './account-mongo-repository'
