+++ tests/infra/db/mongodb/account-mongodb-repository.spec.ts
@@ -1,25 +1,8 @@
-import { mongoHelper } from '../../../../src/infra/db/mongodb';
+import { mongoHelper, AccountMongoRepository } from '../../../../src/infra/db/mongodb';
 import { mockAddAccountParams } from '../../../domain/mocks/mock-account'
-import { CheckAccountByEmailRepository } from '../../../../src/data/protocols'
 import { Collection } from 'mongodb'
+import faker from 'faker';
 
-export class AccountMongoRepository implements CheckAccountByEmailRepository{
-    async checkByEmail(email: string): Promise<CheckAccountByEmailRepository.Result>{
-        accountCollection = mongoHelper.getCollection('accounts') 
-        const account = await accountCollection.findOne({
-            email
-        },{
-            projection: {
-                _id : 1,
-                username : 1,
-                password : 1
-            }
-        })
-        return account !== null
-    }
-
-    
-}
 
 let accountCollection : Collection
 describe('AccountRepository', () => {
@@ -36,12 +19,22 @@ describe('AccountRepository', () => {
         await mongoHelper.disconnect()
     })
 
+    describe('checkByEmail()', () => {
+        
+        test('Should return true if email exist in database', async ()=> {
+            const sut = new AccountMongoRepository();
+            const accountParams = mockAddAccountParams()
+            await accountCollection.insertOne(accountParams)
+            const isValid = await sut.checkByEmail(accountParams.email)
+            expect(isValid).toBe(true)
+        })
 
-    test('Should return true if email exist in database', async ()=> {
-        const sut = new AccountMongoRepository();
-        const accountParams = mockAddAccountParams()
-        await accountCollection.insertOne(accountParams)
-        const isValid = await sut.checkByEmail(accountParams.email)
-        expect(isValid).toBe(true)
+        test('Should return true if email exist in database', async ()=> {
+            const sut = new AccountMongoRepository();
+            const isValid = await sut.checkByEmail(faker.internet.email())
+            expect(isValid).toBe(false)
+        })
+
     })
+   
 })
